backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
google-cloud-firestore>=2.16.0
google-auth>=2.27.0
websockets>=12.0
